Create panel soruce file
